Relocated time_lapse() from Scheduler to main
Also changed output format to the final string
